+++ lib/app/blocs/drive/drive_bloc.dart
@@ -1,0 +1,29 @@
+
+import 'package:bloc/bloc.dart';
+import 'package:googleapis/drive/v3.dart';
+import 'package:http/http.dart' as http;
+import 'package:web/app/models/http_client.dart';
+import 'package:web/app/models/user.dart' as usr;
+import 'package:web/app/blocs/drive/drive_event.dart';
+
+class DriveBloc extends Bloc<DriveEvent, DriveApi> {
+  DriveBloc() : super(null);
+
+  @override
+  Stream<DriveApi> mapEventToState(DriveEvent event) async* {
+    if (event is InitialDriveEvent){
+      yield initialize(event.user);
+    }
+  }
+
+  DriveApi initialize(usr.User user) {
+    http.Client client;
+    if (user != null) {
+      client = ClientWithAuthHeaders(user.headers);
+    } else {
+      client = ClientWithGoogleDriveKey();
+    }
+
+    return DriveApi(client);
+  }
+}
