+++ lib/app/blocs/drive/drive_event.dart
@@ -1,0 +1,10 @@
+import 'package:web/app/models/user.dart' as usr;
+
+abstract class DriveEvent {
+  const DriveEvent();
+}
+
+class InitialDriveEvent extends DriveEvent{
+  final usr.User user;
+  InitialDriveEvent({this.user});
+}
