+++ lib/app/blocs/images/images_event.dart
@@ -1,0 +1,10 @@
+abstract class ImagesEvent {
+  const ImagesEvent();
+}
+
+class GetImagesEvent extends ImagesEvent{
+  final String imageURL;
+
+  GetImagesEvent(this.imageURL);
+
+}
