+++ lib/app/blocs/navigation/navigation_bloc.dart
@@ -1,6 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
-import 'package:web/bloc/navigation/navigation_event.dart';
+import 'package:web/app/blocs/navigation/navigation_event.dart';
+import 'dart:html' as html;
 
 class NavigationBloc extends Bloc<NavigationEvent, dynamic> {
   final GlobalKey<NavigatorState> navigatorKey;
@@ -12,6 +13,12 @@ class NavigationBloc extends Bloc<NavigationEvent, dynamic> {
     switch (event.runtimeType) {
       case NavigatorPopEvent:
         navigatorKey.currentState.pop();
+        break;
+      case NavigateToChangeProfileEvent:
+        html.window.open("https://myaccount.google.com/personal-info", 'Account');
+        break;
+      case NavigateToUpgradeEvent:
+        html.window.open("https://one.google.com/about/plans", 'Upgrade');
         break;
       default:
         navigatorKey.currentState.pushNamed(event.route);
