+++ lib/app/blocs/sharing/sharing_event.dart
@@ -1,0 +1,7 @@
+abstract class ShareEvent {
+  const ShareEvent();
+}
+
+class InitialEvent extends ShareEvent{}
+class StartSharingEvent extends ShareEvent{}
+class StopSharingEvent extends ShareEvent{}
