--- lib/bloc/navigation/navigation_bloc_test.dart
@@ -1,7 +1,0 @@
-import 'package:bloc/bloc.dart';
-import 'package:bloc_test/bloc_test.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:web/bloc/navigation/navigation_bloc.dart';
-import 'package:web/bloc/navigation/navigation_event.dart';
-
-main() {}
