+++ lib/main.dart
@@ -2,17 +2,16 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:web/bloc/authentication/authentication_bloc.dart';
-import 'package:web/bloc/authentication/authentication_event.dart';
-import 'package:web/bloc/navigation/navigation_bloc.dart';
-import 'package:web/locator.dart';
+import 'package:web/app/blocs/authentication/authentication_bloc.dart';
+import 'package:web/app/blocs/authentication/authentication_event.dart';
+import 'package:web/app/blocs/drive/drive_bloc.dart';
+import 'package:web/app/blocs/navigation/navigation_bloc.dart';
 import 'package:web/route.dart';
-import 'package:web/theme.dart';
 import 'package:web/ui/pages/static/home.dart';
+import 'package:web/ui/theme/theme.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  setupLocator();
   runApp(MyApp());
 }
 
@@ -25,10 +24,12 @@ class _MyAppState extends State<MyApp> {
   final GlobalKey<NavigatorState> _navigatorKey = GlobalKey();
   AuthenticationBloc _authenticationBloc;
   NavigationBloc _navigationBloc;
+  DriveBloc _driveBloc;
 
   @override
   void initState() {
     super.initState();
+    _driveBloc = DriveBloc();
     _navigationBloc = NavigationBloc(navigatorKey: _navigatorKey);
     _authenticationBloc = AuthenticationBloc();
     _authenticationBloc.add(AuthenticationSilentSignInEvent());
@@ -39,12 +40,16 @@ class _MyAppState extends State<MyApp> {
     super.dispose();
     _navigationBloc.close();
     _authenticationBloc.close();
+    _driveBloc.close();
   }
 
   @override
   Widget build(BuildContext context) {
     return MultiBlocProvider(
       providers: [
+        BlocProvider<DriveBloc>(
+          create: (BuildContext context) => _driveBloc,
+        ),
         BlocProvider<NavigationBloc>(
           create: (BuildContext context) => _navigationBloc,
         ),
