+++ pubspec.yaml
@@ -30,7 +30,6 @@ dependencies:
   html: ^0.14.0+4
   flutter_spinkit: ^4.1.2+1
   responsive_builder: ^0.1.2
-  get_it: ^5.0.1
   google_sign_in: ^4.5.4
   provider: ^4.3.2+2
   cached_network_image: ^2.3.2+1
